aw-pywebview/pyproject.toml
[tool.poetry]
name = "aw-pywebview"
version = "0.1.0"
description = "PyWebView shell for ActivityWatch (all-in-one)"
authors = ["ActivityWatch Developers"]
license = "MPL-2.0"

[tool.poetry.scripts]
aw-pywebview = "aw_pywebview:main"
aw-pywebview-report = "aw_pywebview.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
pywebview = "^5.1"
requests = "^2.32.3"
Pillow = "^10.4.0"
aw-core = {path = "../aw-core"}
aw-client = {path = "../aw-client"}
aw-server = {path = "../aw-server"}
aw-watcher-afk = {path = "../aw-watcher-afk"}
aw-watcher-window = {path = "../aw-watcher-window"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
